requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiomysql==0.2.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
asyncio==3.4.3
attrs==25.3.0
bcrypt==4.2.2
beautifulsoup4==4.13.4
binencrypt==2.5
bitarray==2.9.3
bitstring==4.2.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
cryptodo==3.3
cryptography==41.0.7
distro==1.9.0
ecdsa==0.19.0
email-validator==2.2.0
esptool==4.8.1
et_xmlfile==2.0.0
Faker==37.4.2
fastapi==0.104.1
flatbuffers==24.3.25
frozenlist==1.7.0
gast==0.6.0
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.40.3
google-cloud-core==2.5.0
google-cloud-documentai==3.9.0
google-cloud-storage==3.8.0
google-cloud-vision==3.4.4
google-crc32c==1.8.0
google-generativeai==0.3.2
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.68.1
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
intelhex==2.3.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.7.0
krishto==0.2
langchain==1.2.0
langchain-core==1.2.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.6.3
namex==0.0.8
numpy==1.26.4
openai==2.9.0
opencv-python==4.10.0.84
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.11.5
ormsgpack==1.12.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.1.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
Pygments==2.18.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyMySQL==1.1.0
pypdfium2==4.30.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
reedsolo==1.7.0
reportlab==4.4.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
supabase==2.11.1
sympy==1.13.3
tabula-py==2.10.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
user-agent==0.1.10
uuid_utils==0.12.0
uv==0.9.18
uvicorn==0.24.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.3.2
xxhash==3.6.0
yarl==1.20.1
yt-dlp==2025.10.14
zstandard==0.25.0
